moved traits out of schema
